EDSC-3929: add tests
